requirements.txt
Django==1.9
pytz==2015.7
pycrypto==2.6.1
libcnml==0.9.2
django-smart-selects==1.2.2
